// src/app/data/enums/SectionKey.ts
////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

export enum SectionKey {
  AboutMe = "about-me",
  Projects = "projects",
  Settings = "settings"
}

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////